Description
Allow option in whether or not to create a route to the NAT gateway in private subnets.

Motivation and Context

There are use cases where the creation of this route is undesired.  One example is in the use of a VPC network firewall, where the route destination would be the VPC endpoint of the firewall.

Breaking Changes

None